refactor: init types crate
