+++ lib/Cargo.toml
@@ -13,16 +13,14 @@ protox = { workspace = true }
 tonic-prost-build = { workspace = true }
 
 [dependencies]
-anyhow = { workspace = true, features = ["backtrace"] }
+anyhow = { workspace = true }
 async-lock = { workspace = true }
 bincode = { workspace = true }
-bitcoin = { workspace = true, features = ["serde"] }
-blake3 = { workspace = true }
-borsh = { workspace = true, features = ["derive"] }
+bitcoin = { workspace = true }
+borsh = { workspace = true }
 byteorder = { workspace = true }
 bytes = { workspace = true }
-clap = { workspace = true, features = ["derive"], optional = true }
-ed25519-dalek = { workspace = true, features = ["batch", "serde"] }
+ed25519-dalek = { workspace = true }
 ed25519-dalek-bip32 = { workspace = true }
 educe = { workspace = true, features = ["Clone", "Debug"] }
 error-fatality = { workspace = true }
@@ -31,9 +29,8 @@ futures = { workspace = true, features = ["alloc"] }
 governor = { workspace = true, features = ["quanta", "std"] }
 hashlink = { workspace = true, features = ["serde_impl"] }
 heed = { workspace = true }
-hex = { workspace = true, features = ["serde"] }
+hex = { workspace = true }
 jsonrpsee = { workspace = true }
-merkle-cbt = { workspace = true }
 nonempty = { workspace = true, features = ["serialize"] }
 parking_lot = { workspace = true }
 prost = { workspace = true }
@@ -42,15 +39,12 @@ quinn = { workspace = true }
 rayon = { workspace = true }
 rcgen = { workspace = true }
 rustls = { workspace = true, features = ["ring"] }
-rustreexo = { workspace = true, features = ["with-serde"] }
-smallvec = { workspace = true, features = ["write"] }
-semver = { workspace = true, features = ["serde"] }
+semver = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
-serde_with = { workspace = true }
 sneed = { workspace = true, features = ["observe"] }
-strum = { workspace = true }
 thiserror = { workspace = true }
+thunder_types = { path = "../types", features = ["heed"] }
 tiny-bip39 = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread", "sync"] }
 tokio-stream = { workspace = true, features = ["sync"] }
@@ -59,13 +53,13 @@ tonic = { workspace = true }
 tonic-prost = { workspace = true }
 tracing = { workspace = true }
 transitive = { workspace = true }
-utoipa = { workspace = true, features = ["macros", "non_strict_integers"] }
 
 [dev-dependencies]
+hashlink = { workspace = true }
 temp-dir = { workspace = true }
 
 [features]
-clap = ["dep:clap"]
+clap = ["thunder_types/clap"]
 
 [lints]
 workspace = true
