+++ types/wallet.rs
@@ -1,0 +1,16 @@
+use bitcoin::Amount;
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize, ToSchema)]
+pub struct Balance {
+    #[serde(rename = "total_sats", with = "bitcoin::amount::serde::as_sat")]
+    #[schema(value_type = u64)]
+    pub total: Amount,
+    #[serde(
+        rename = "available_sats",
+        with = "bitcoin::amount::serde::as_sat"
+    )]
+    #[schema(value_type = u64)]
+    pub available: Amount,
+}
